add detection for blocking channels and ensure goroutine dies
stream.stream has a blocking condition when trying to push onto the
error channel. this channel can get closed by other goroutines and
will cause this goroutine to leak.

by passing to a default case we can drop the error message when we are
unable to return it and move to reconnect logic.

We also add in detection for stream cancellation from higher calling
goroutines in preparation for shutdown logic to be added